modified process timer example

+++ child_process/child_process.cpp
@@ -67,9 +67,10 @@ public:
         ProcessExitCallback callback = std::bind(&ProcessRunner::exitCallback, shared_from_this(), std::placeholders::_1, std::placeholders::_2);
         asyncHandler_ = std::make_unique<AptAsyncHandler>(callback);
         childProcess_ = std::make_unique<boost::process::child>(cmd, *asyncHandler_, boost::process::std_out > boost::process::null, boost::process::std_err > boost::process::null, io_context_);
+        timestamp_ = std::chrono::steady_clock::now();
 
         auto instance = shared_from_this();
-        processTimer_ = std::make_unique<boost::asio::steady_timer>(io_context_, std::chrono::seconds(5));
+        processTimer_ = std::make_unique<boost::asio::steady_timer>(io_context_, std::chrono::seconds(120));
         std::cout << "starting timer" << std::endl;
         processTimer_->async_wait(
             [instance](boost::system::error_code const& timer_ec)
@@ -81,11 +82,21 @@ public:
                 else
                 {
                     std::cout << "timer expired";
-                    if (instance->childProcess_)
+                    auto elapsed = std::chrono::steady_clock::now() - instance->timestamp_;
+                    std::cout << " (" << std::chrono::duration_cast<std::chrono::seconds>(elapsed).count() << " seconds elapsed)";
+                    if (elapsed > std::chrono::minutes(2))
                     {
-                        std::cout << " -- killing child process";
-                        std::error_code processEc;
-                        instance->childProcess_->terminate(processEc);
+                        if (instance->childProcess_)
+                        {
+                            std::cout << " -- killing child process";
+                            std::error_code processEc;
+                            instance->childProcess_->terminate(processEc);
+                        }
+                    }
+                    else
+                    {
+                        auto remaining = std::chrono::minutes(2) - elapsed;
+                        std::cout << " -- not enough time passed (" << std::chrono::duration_cast<std::chrono::seconds>(remaining + std::chrono::seconds(5)).count() << " remaining)";
                     }
                     std::cout << std::endl;
                 }
@@ -110,8 +121,9 @@ public:
     }
 
 private:
-    const std::string cmd{"/usr/bin/sleep 10"};
+    const std::string cmd{"/usr/bin/sleep 125"};
     boost::asio::io_context io_context_;
+    std::chrono::steady_clock::time_point timestamp_;
     std::unique_ptr<boost::process::child> childProcess_;
     std::unique_ptr<AptAsyncHandler> asyncHandler_;
     std::unique_ptr<boost::asio::steady_timer> processTimer_;
